Enable all warnings

+++ src/math/poly.cc
@@ -37,7 +37,7 @@ void polyfit(const cv::Mat& src_x, const cv::Mat& src_y, cv::Mat& dst,
 std::vector<float> polyfit(const std::vector<cv::Point>& src, int order) {
   cv::Mat src_x(src.size(), 1, CV_32F);
   cv::Mat src_y(src.size(), 1, CV_32F);
-  for (int i = 0; i < src.size(); i++) {
+  for (size_t i = 0; i < src.size(); i++) {
     src_x.at<float>(i, 0) = static_cast<float>(src[i].x);
     src_y.at<float>(i, 0) = static_cast<float>(src[i].y);
   }
